refactor(pages): product and thank you moved into pages folder
