+++ .prettierrc
@@ -1,0 +1,1 @@
+{ "plugins": ["prettier-plugin-tailwindcss"] }
