+++ src/App.tsx
@@ -1,12 +1,13 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { ProductMaster } from "./Product";
-import { ThankYouPage } from "./ThankYouPage";
-
+import { ProductMaster } from "./pages/Product";
+import { ThankYouPage } from "./pages/ThankYouPage";
+import Home from "./pages/Home";
 const App = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<ProductMaster />} />
+        <Route path="/" element={<Home />} />
+        <Route path="/product" element={<ProductMaster />} />
         <Route path="/thank-you" element={<ThankYouPage />} />
       </Routes>
     </BrowserRouter>
