+++ src/pages/Home.tsx
@@ -1,0 +1,9 @@
+const Home = () => {
+  return (
+    <div className="flex bg-red-200">
+      <h2>Home</h2>
+    </div>
+  );
+};
+
+export default Home;
