+++ vite.config.ts
@@ -1,13 +1,12 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
 
-// https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
   define: {
-    'process.env': {},
+    "process.env": {},
   },
-    server: {
+  server: {
     port: 3001,
   },
-})
+});
